requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.9
djangorestframework==3.12.4
idna==3.3
python-decouple==3.5
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
typing_extensions==4.0.0
urllib3==1.26.7
